# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API document for NMC bookstore e-commerce website",
        "title": "NMC Bookstore",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Nguyen Minh Chien",
            "url": "https://github.com/Chien179",
            "email": "chiennm456@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/admin/subgenres": {
            "post": {
                "description": "Use this API to create subgenre",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subgenres"
                ],
                "summary": "Create subgenre",
                "parameters": [
                    {
                        "description": "Create subgenre",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createSubgenreRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.Subgenre"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/admin/subgenres/delete/{id}": {
            "delete": {
                "description": "Use this API to delete subgenre",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subgenres"
                ],
                "summary": "Delete subgenre",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "delete subgenre",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/admin/subgenres/update/{id}": {
            "put": {
                "description": "Use this API to update subgenre",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subgenres"
                ],
                "summary": "Update subgenre",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Update subgenre id",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update subgenre data",
                        "name": "Request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/api.updateSubgenreData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.Subgenre"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/subgenres/{genre_id}": {
            "get": {
                "description": "Use this API to list subgenre",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subgenres"
                ],
                "summary": "List subgenre",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "list subgenre",
                        "name": "genre_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.Subgenre"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.createSubgenreRequest": {
            "type": "object",
            "required": [
                "genre_id",
                "name"
            ],
            "properties": {
                "genre_id": {
                    "type": "integer",
                    "minimum": 1
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "api.updateSubgenreData": {
            "type": "object",
            "properties": {
                "genre_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "db.Subgenre": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "genres_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
...
}